payment/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import viewsets

from payment.filters import PaymentFilter
from payment.models import Payment
from payment.serializers import PaymentSerializer


class PaymentViewSet(viewsets.ModelViewSet):
    queryset = Payment.objects.all()
    serializer_class = PaymentSerializer
    filter_backends = [DjangoFilterBackend]
    filterset_class = PaymentFilter
