requirements.txt
asgiref
Django
djangorestframework
Pillow
pip
psycopg2-binary
pytz
setuptools
sqlparse
tzdata
wheel
django-filter
